# ENH, MAINT: simpler calculate_areas()